make separate utils file
